% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{active_return}
\alias{active_return}
\title{Active return}
\usage{
active_return(x, y, t = 252, method = "geometric")
}
\arguments{
\item{x}{asset returns}

\item{y}{benchmark returns}

\item{t}{frequency of data. 1: yearly, 4: quarterly, 12: monthly, 52: weekly, 252: daily}

\item{method}{"geometric" or "simple" (def: "geometric")}
}
\value{
numeric
}
\description{
Asset annualized return minus Benchmark annualized return
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,0.014,0.024,0.015,0.066,-0.014,0.039)
yy <- c(0.04,-0.022,0.043,0.028,-0.078,-0.011,0.033,-0.049,0.09,0.087)
active_return(xx,yy,t=12)

}
