% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{ann_risk}
\alias{ann_risk}
\title{Annualized Risk}
\usage{
ann_risk(x, t = 252)
}
\arguments{
\item{x}{asset returns}

\item{t}{frequency of data. 1: yearly, 4: quarterly, 12: monthly, 52: weekly, 252: daily}
}
\value{
numeric
}
\description{
Annualized standard deviation
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,0.014,0.024,0.015,0.066,-0.014,0.039)
ann_risk(xx,t=12)

}
