% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{ann_sharpe_ratio}
\alias{ann_sharpe_ratio}
\title{Annualized Sharpe ratio}
\usage{
ann_sharpe_ratio(x, frisk = 0, t = 252, is_geom = TRUE)
}
\arguments{
\item{x}{asset returns}

\item{frisk}{annual free-risk rate. For daily rate (frisk/252), weekly (frisk/52), monthly (frisk/12), etc}

\item{t}{frequency of data. 1: yearly, 4: quarterly, 12: monthly, 52: weekly, 252: daily}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
numeric
}
\description{
Sharpe ratio calculated on annualized excess return / annualized standard deviation
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,0.014,0.024,0.015,0.066,-0.014,0.039)
ann_sharpe_ratio(xx,frisk=0.02/12,t=12)

}
