% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{ann_sortino_ratio}
\alias{ann_sortino_ratio}
\title{Annualized Sortino ratio}
\usage{
ann_sortino_ratio(x, t = 252, ann_mar = 0, is_geom = TRUE)
}
\arguments{
\item{x}{asset returns}

\item{t}{frequency of data. 1: yearly, 4: quarterly, 12: monthly, 52: weekly, 252: daily}

\item{ann_mar}{annual minimum acceptable return}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
numeric
}
\description{
Annualized sortino ratio
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
ann_sortino_ratio(xx,t=12,ann_mar=0.01)

}
