% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_xts}
\alias{as_xts}
\title{As xts}
\usage{
as_xts(x)
}
\arguments{
\item{x}{vector/matrix/data.frame/... object}
}
\value{
xts
}
\description{
Convert vector/matrix/data.frame to xts object. Row names are dates with backward calculation from today.
}
\examples{
as_xts(matrix(NA,10,2))
as_xts(stats::rnorm(10))
as_xts(data.frame(matrix(NA,10,2)))

}
