% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{avg_drawdown}
\alias{avg_drawdown}
\title{Average Drawdown}
\usage{
avg_drawdown(x, is_geom = TRUE)
}
\arguments{
\item{x}{asset returns}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
numeric
}
\description{
Return the average drawdown from asset returns
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
avg_drawdown(xx)

}
