% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{burke_ratio}
\alias{burke_ratio}
\title{Burke ratio}
\usage{
burke_ratio(
  x,
  ann_frisk = 0,
  t = 252,
  is_geom = TRUE,
  method = c("simple", "modified")
)
}
\arguments{
\item{x}{asset returns}

\item{ann_frisk}{annual free risk}

\item{t}{frequency of data. 1: yearly, 4: quarterly, 12: monthly, 52: weekly, 252: daily}

\item{is_geom}{TRUE geometric, FALSE simple}

\item{method}{"simple" or "modified" (def: "simple")}
}
\value{
numeric
}
\description{
A risk-adjusted measure with free risk and drawdowns. For the 'simple' mode the excess return over free risk is divided by the square root of
the sum of the square of the drawdowns. For the 'modified' mode the Burke Ratio is multiplied
by the square root of the number of data.
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
burke_ratio(xx,ann_frisk=0.01,t=12)
burke_ratio(xx,ann_frisk=0.01,t=12,method="modified")

}
