% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{calmar_ratio}
\alias{calmar_ratio}
\title{Calmar ratio}
\usage{
calmar_ratio(x, ann_frisk = 0, t = 252, is_geom = TRUE)
}
\arguments{
\item{x}{asset returns}

\item{ann_frisk}{annual free risk value}

\item{t}{frequency of data. 1: yearly, 4: quarterly, 12: monthly, 52: weekly, 252: daily}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
numeric
}
\description{
A risk-adjusted measure like Sharpe ratio that uses maximum drawdown instead of standard deviation for risk
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
calmar_ratio(xx,ann_frisk = 0.01,t=12)

}
