% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{downside_potential}
\alias{downside_potential}
\title{Downside potential}
\usage{
downside_potential(x, mar = 0)
}
\arguments{
\item{x}{asset returns}

\item{mar}{minimum acceptable return}
}
\value{
numeric
}
\description{
Downside potential is the first lower partial moment
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
downside_potential(xx)

}
