% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{downside_risk}
\alias{downside_risk}
\title{Downside risk}
\usage{
downside_risk(x, mar = 0)
}
\arguments{
\item{x}{asset returns}

\item{mar}{minimum acceptable return}
}
\value{
numeric
}
\description{
Downside Risk or Semi-Standard Deviation.Measures  the  variability  of  underperformance  below  a  minimum  target   rate
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
downside_risk(xx,mar=0.1/100)

}
