% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{drawdown}
\alias{drawdown}
\title{Drawdown}
\usage{
drawdown(x, is_geom = TRUE)
}
\arguments{
\item{x}{asset returs}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
numeric vector of length x
}
\description{
Calculate drawdown array from asset returns. Return a vector with numbers >= 0
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
drawdown(xx)

}
