% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{drawdown_info}
\alias{drawdown_info}
\title{Drawdown information}
\usage{
drawdown_info(x, is_geom = TRUE)
}
\arguments{
\item{x}{asset returns}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
list
}
\description{
Return a list of drawdown info: value, start index, trough index, end index, duration, peak-to-trough, recovery period
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
drawdown_info(xx)

}
