% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{excess_return}
\alias{excess_return}
\title{Excess return}
\usage{
excess_return(x, frisk = 0)
}
\arguments{
\item{x}{asset returns}

\item{frisk}{annual free-risk rate (def: 0). For daily rate (frisk/252), weekly (frisk/52), monthly (frisk/12), etc}
}
\value{
numeric
}
\description{
Return on asset - risk free rate
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,0.014,0.024,0.015,0.066,-0.014,0.039)
excess_return(xx,frisk=0.02/12)

}
