% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric
}
\description{
Kurtosis
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
kurtosis(xx)

}
