% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{martin_ratio}
\alias{martin_ratio}
\title{Martin ratio}
\usage{
martin_ratio(x, ann_frisk = 0, t = 252, is_geom = TRUE)
}
\arguments{
\item{x}{asset returns}

\item{ann_frisk}{annual free risk}

\item{t}{frequency of data. 1: yearly, 4: quarterly, 12: monthly, 52: weekly, 252: daily}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
numeric
}
\description{
A risk-adjusted measure with free risk and Ulcer index.
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
martin_ratio(xx,ann_frisk=0.01,t=12)

}
