% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{norminv}
\alias{norminv}
\title{Inverse of the normal cumulative distribution function (cdf)}
\usage{
norminv(p, mu = 0, sigma = 1)
}
\arguments{
\item{p}{probability value in range 0-1}

\item{mu}{mean value}

\item{sigma}{standard deviation}
}
\value{
numeric
}
\description{
Returns the inverse cdf for the normal distribution with mean MU and standard deviation SIGMA at P value
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
norminv(0.01,mean(xx),sd(xx))

}
