% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{normpdf}
\alias{normpdf}
\title{Normal probability density function (pdf)}
\usage{
normpdf(x, mu = 0, sigma = 1)
}
\arguments{
\item{x}{real value}

\item{mu}{mean value}

\item{sigma}{standard deviation}
}
\value{
numeric
}
\description{
Returns the pdf of the normal distribution with mean MU and standard deviation SIGMA
evaluated at the values in X
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
normpdf(0.01,mean(xx),sd(xx))

}
