% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{omega_ratio}
\alias{omega_ratio}
\title{Omega ratio}
\usage{
omega_ratio(x, mar = 0)
}
\arguments{
\item{x}{asset returns}

\item{mar}{minimum acceptable return}
}
\value{
numeric
}
\description{
A gain/loss ratio: upside potential/downside potential
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
omega_ratio(xx)

}
