% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{param_cvar}
\alias{param_cvar}
\title{Parametric Conditional Value At Risk}
\usage{
param_cvar(mu = 0, sigma = 1, p = 0.95)
}
\arguments{
\item{mu}{mean value}

\item{sigma}{standard deviation}

\item{p}{confidence level}
}
\value{
numeric
}
\description{
Parametric Conditional Value-At-Risk. More sensitive to the shape of the loss distribution in the tails.
Also known as Expected Shortfall (ES), Expected Tail Loss (ETL).
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
param_cvar(mean(xx),sd(xx))

}
