% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{param_var}
\alias{param_var}
\title{Parametric Value At Risk}
\usage{
param_var(x, p = 0.95)
}
\arguments{
\item{x}{asset returns}

\item{p}{confidence level}
}
\value{
numeric
}
\description{
Parametric or gaussian value at risk
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
param_var(xx,p=0.95)

}
