% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{rand_ts}
\alias{rand_ts}
\title{Random time series}
\usage{
rand_ts(m = 20, n = 1, as_ret = F)
}
\arguments{
\item{m}{num of samples}

\item{n}{num of series}

\item{as_ret}{TRUE for returns, FALSE for values}
}
\value{
matrix
}
\description{
Create random time series with normal distribution
}
\examples{
rand_ts(20,2)
rand_ts(20,3,as_ret=TRUE)

}
