% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric
}
\description{
Skewness
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
skewness(xx)
# 0.6174813
}
