% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{sortino_ratio}
\alias{sortino_ratio}
\title{Sortino ratio}
\usage{
sortino_ratio(x, mar = 0)
}
\arguments{
\item{x}{asset returns}

\item{mar}{minimum acceptable return}
}
\value{
numeric
}
\description{
Sortino ratio
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
sortino_ratio(xx)

}
