% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{ulcer_index}
\alias{ulcer_index}
\title{Ulcer index}
\usage{
ulcer_index(x, is_geom = TRUE)
}
\arguments{
\item{x}{asset returns}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
numeric
}
\description{
RUlcer Index of Peter G. Martin (1987). The impact of long, deep drawdowns will have significant impact because the underperformance since the last peak is squared
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
ulcer_index(xx)

}
