% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{upside_potential}
\alias{upside_potential}
\title{Upside potential}
\usage{
upside_potential(x, mar = 0)
}
\arguments{
\item{x}{asset returns}

\item{mar}{minimum acceptable return}
}
\value{
numeric
}
\description{
Upside potential
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
upside_potential(xx,mar=0.1/100)

}
