% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{upside_risk}
\alias{upside_risk}
\title{Upside risk}
\usage{
upside_risk(x, mar = 0)
}
\arguments{
\item{x}{asset returns}

\item{mar}{minimum acceptable return}
}
\value{
numeric
}
\description{
Measures  the  variability  of  overperformance  above  a  minimum  target   rate
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,-0.014,-0.024,0.015,0.066,-0.014,0.039)
upside_risk(xx,mar=0.1/100)

}
