% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{windowing}
\alias{windowing}
\title{Windowing}
\usage{
windowing(
  x = NULL,
  k = NULL,
  s = 1,
  method = c("window", "anchored"),
  overlapping = F
)
}
\arguments{
\item{x}{number of elements to split as scalar}

\item{k}{windows size (training)}

\item{s}{sliding step (test)}

\item{method}{"window" (default) or "anchored" (no old data forgetting)}

\item{overlapping}{TRUE, skip training+test windows (e.g. training 1:12/test 13:15,training 16:27/test 28:30,...)}
}
\value{
list
}
\description{
Create a training/test set with sliding windows. Return a list of indices for training and test set
}
\examples{
windowing(60,12,3)
windowing(60,12,3,"anchored")
windowing(60,12,3,overlapping=TRUE)

}
