% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quants.R
\name{ann_return}
\alias{ann_return}
\title{Annualized return}
\usage{
ann_return(x, t = 252, is_geom = TRUE)
}
\arguments{
\item{x}{asset returns}

\item{t}{frequency of data. 1: yearly, 4: quarterly, 12: monthly, 52: weekly, 252: daily}

\item{is_geom}{TRUE geometric, FALSE simple}
}
\value{
numeric
}
\description{
Average annualized returns
}
\examples{
xx <- c(0.003,0.026,0.015,-0.009,0.014,0.024,0.015,0.066,-0.014,0.039)
ann_return(xx,t=12)

}
